fix codecheck bug



# message auto-generated for no-merge-commit merge:
!181 merge master into master

fix codecheck bug

Created-by: guijianwei
Commit-by: guijianwei
Merged-by: turing_project1
Description: ## 描述
<!--在这里详细描述你的改动，包括改动的原因和所采取的方法。-->

## 关联的Issue
<!-- 如果这个PR是为了解决特定的Issue，请在这里提供Issue链接。例如：关联Issue #123-->
<!-- 如果这个PR是为了解决特定的问题单，请在这里描述问题单单号。-->

## 测试
<!--描述进行了哪些测试来验证你的改动。包括但不限于二级冒烟、算子泛化等。-->
![image.png](https://raw.gitcode.com/user-images/assets/7665709/bbcdf00a-be63-48a8-887c-25d8b24b2a70/image.png 'image.png')
![image.png](https://raw.gitcode.com/user-images/assets/7665709/0f529c6f-69ae-4d1b-a7cd-24a274df7c4f/image.png 'image.png')
![image.png](https://raw.gitcode.com/user-images/assets/7665709/aa5354d0-291e-4135-b093-12a7156d44cb/image.png 'image.png')
![image.png](https://raw.gitcode.com/user-images/assets/7665709/e637ef06-52fa-4bfa-a295-dc6580f449b4/image.png 'image.png')
![image.png](https://raw.gitcode.com/user-images/assets/7665709/762f98e9-ba34-4bbe-9687-9622834ac103/image.png 'image.png')
## 文档更新
<!--如果这个PR包含文档的更新，请在这里指出。例如：更新了README.md文件。-->

## 类型标签
<!--  [x] 表示选中 -->
- [ ] Bug修复
- [ ] 新特性
- [ ] 性能优化
- [ ] 文档更新
- [ ] 其他，请描述：




See merge request: cann/ops-nn!181

+++ activation/selu_grad/examples/test_aclnn_selu_grad.cpp
@@ -86,8 +86,8 @@ int main() {
   aclTensor* gradOutput = nullptr;
   aclTensor* gradInput = nullptr;
   std::vector<float> selfHostData = {0, 1, 2, 3, 4, 5, 6, 7};
-  std::vector<int> gradOutputHostData = {1, 1, 1, 1, 1, 1, 1, 1};
-  std::vector<int> gradInputHostData = {0, 0, 0, 0, 0, 0, 0, 0};
+  std::vector<float> gradOutputHostData = {1, 1, 1, 1, 1, 1, 1, 1};
+  std::vector<float> gradInputHostData = {0, 0, 0, 0, 0, 0, 0, 0};
 
   ret = CreateAclTensor(selfHostData, selfShape, &selfDeviceAddr, aclDataType::ACL_FLOAT, &self);
   CHECK_RET(ret == ACL_SUCCESS, return ret);
